requirements.txt
numpy~=2.2.2
scipy~=1.15.1
networkx~=3.4.2
matplotlib~=3.10.0
click~=8.1.8
python-dotenv~=1.0.1
keras~=3.8.0
scikit-learn~=1.6.1
pandas~=2.2.3
setuptools~=75.8.0
scikit-image~=0.25.1
tensorflow~=2.18

tqdm~=4.67.1
keras_hub~=0.18.1
